fix: rename to shim.js
